Database incorporated, for faster retrieval
Using JDBC and MySQL, a database was built to hold SFD values of each
color of each image in the images folder. The values need not be
computed every time the program is run - If present in the database, the
cost of computation is just a database access, and if absent in the
database, the values are calculated and automatically inserted for
future reference.

This system works much more faster than before, on my desktop...However,
I fail to see any difference between the performance of the old system
and new one on my netbook which has a 1GHz AMD Processor, which leads me
to believe that this system cannot be used on low-end computers.

That's about it.....